+++ Sources/KeyboardKit/Views/System/SystemKeyboardButtonRowItem.swift
@@ -108,13 +108,24 @@ public extension View {
      */
     @ViewBuilder
     func rowItemWidth(for item: KeyboardLayoutItem, totalWidth: CGFloat, referenceWidth: CGFloat) -> some View {
+        if let width = rowItemWidthValue(for: item, totalWidth: totalWidth, referenceWidth: referenceWidth), width > 0 {
+            self.frame(width: width)
+        } else {
+            self.frame(maxWidth: .infinity)
+        }
+    }
+}
+
+private extension View {
+
+    func rowItemWidthValue(for item: KeyboardLayoutItem, totalWidth: Double, referenceWidth: Double) -> Double? {
         let insets = item.insets.leading + item.insets.trailing
         switch item.size.width {
-        case .available: self.frame(maxWidth: .infinity)
-        case .input: self.frame(width: referenceWidth - insets)
-        case .inputPercentage(let percent): self.frame(width: percent * referenceWidth - insets)
-        case .percentage(let percent): self.frame(width: percent * totalWidth - insets)
-        case .points(let points): self.frame(width: points - insets)
+        case .available: return nil
+        case .input: return referenceWidth - insets
+        case .inputPercentage(let percent): return percent * referenceWidth - insets
+        case .percentage(let percent): return percent * totalWidth - insets
+        case .points(let points): return points - insets
         }
     }
 }