+++ Tests/LinuxMain.swift
@@ -1,7 +1,0 @@
-import XCTest
-
-import KeyboardKitTests
-
-var tests = [XCTestCaseEntry]()
-tests += KeyboardKitTests.allTests()
-XCTMain(tests)